:boom: Don't limit onClick on `BodyRow` children based on their parentNodes
event.stopPropagation should be called on children down the tree that have `onClick` events. We _could_ try to detect if the current target component has an onClick prop and then wrap that with a function that calls stopPropagation, but I believe this abstraction doesn't belong in @teamleader/ui. It would give more issues than it solves.