Make stepper events trigger the onChange event
I did this so that the stepper can be used when
passing native value and onChange props to
the NumericInput

For example a form library that passes value and onChange
to this component will now be able to work with the
stepper out of the box.

Current behavior is still supported (onChange(event, value)).