Add scrollButtonOffset
to allow the scroll elements to be placed
in a way the entire TabGroup uses full width of
the parent container.

When no scroll elements are visible the TabGroup respects
the padding given by the parent container

The attribute is optional and defaults to old behaviour
so there is no breaking change involved.

(Discussed with Joren Devocht and Stephen Verhalleman)