:bug: fix mouse down - drag motion causing the overlay to be closed
In Chrome, Safari & Edge the click target is the target that was last hovered before releasing the mouse button. Only Firefox correctly detects the target the original target where the click began. So in order to fix this, we need to keep track of the original click target and use that when comparing.